Re-implemented examples using a shared compiler class.
Spec definition is triggered when SenTestCase requests invocations for testing.
The functions it(), example() and specify() add examples to the shared compiler.
The shared compiler will be responsible for 'compiling' the blocks with the correct hooks in place, but for now simply returns its examples.

BHVExampleInvocation (a private class to BHVSpec) maintains a reference to an example and overrides -invoke to perform -execute on its example, executing the test.

BHVSpec -name is overridden to provide for prettier output, and is tested at a higher level using Cucumber.