Rewritten specs.
Not really sure how to write specs for a testing framework, using the same framework. It's very difficult.

We have two choices: write tests using OCUnit and then upgrade them to use Beehive at a later date, or ignore tests for now and come back to it later.

I vote for OCUnit.