Made significant changes; see full comments.
Rather than having a shared suite, we're using a suite for each spec. BHVSuiteRegistry has been added for this purpose – suites are registered for a given class, then retrieved at any time.

We'll be using block variables to define it(), describe(), pending(), etc. This is in order to re-define these variables when adding examples to a context other than the suite.

The suite is itself a context, and may or may not inherit from it.