Re-added 'it' for defining examples.
This time based on suite locking - examples are added to every unlocked suite.
This is as opposed to the general solution of keeping track of the current spec that is being compiled through NSThread or otherwise.
We just use one singleton and no more- BHVSuiteRegistry.

BHVExample now has a -name attribute rather than using its description.