Added context function.
Need to compile examples in contexts down and concatenate their names.
Also need to add separated tests for adding examples and contexts to locked / unlocked suites - that is, that they do not add to locked suites and do to unlocked.

May also fix the copying issue? If a context is handed to multiple suites, its examples are re-added over and over. Actually, could just blank the context's examples each time and re-build them? This makes the most sense, IMO.