Moved DSL functions into one file.
Project was getting too large with one-file-per-function, which was really just a growth from using one for `it()` and not really managing it. Now all functions are in BHVDSL, which is also a class. The class provides private class methods used by the functions, rather than keeping the implementation in functions. You never know, it might need to be stubbed etc, which can't be done with functions.