Updated README.
It has become apparent that, due to the limitations of Objective-C as a compiled language, 'should' cannot be used in the context of both with a in-line and otherwise subject. It must only be inline. If it ever becomes possible to do so, it will be placed back on the roadmap.