Switched test focus to +testInvocations and -name.
They are the most important functions of the specification unit. This change simplified things greatly – the tests for adding examples and hooks depended on direct access to the class methods +contexts, +examples, and +hooks, but these methods should not need to be made available to the test suite as they are not the primary focus of the tests. The goal of +addExample etc is to make the example available for the creation of an invocation.