Added DSL.
Modified +testInvocations tests to make use of subclasses instead of constructing example structures in each test. This is more awkward, but works from the top of the API (how the client will be using the code) as it should, imo. If we didn't, the number of tests would be astronomical.